v1.4 XW 4-11-2017 cleaned up the code

+++ libraries/KalmanFilter-master/Kalman.cpp
@@ -19,21 +19,21 @@
 
 Kalman::Kalman() {
     /* We will set the variables like so, these can also be tuned by the user */
-    Q = 0.001f;
-    Q_bias = 0.003f;
-    R_measure = 0.03f;
+    Q = 0.001;
+    Q_bias = 0.003;
+    R_measure = 0.03;
 
-    x = 0.0f; // Reset x
-    bias = 0.0f; // Reset bias
+    x = 0; // Reset x
+    bias = 0; // Reset bias
 
-    P[0][0] = 0.0f; // Since we assume that the bias is 0 and we know the starting angle (use setAngle), the error covariance matrix is set like so - see: http://en.wikipedia.org/wiki/Kalman_filter#Example_application.2C_technical
-    P[0][1] = 0.0f;
-    P[1][0] = 0.0f;
-    P[1][1] = 0.0f;
+    P[0][0] = 0; // Since we assume that the bias is 0 and we know the starting angle (use setAngle), the error covariance matrix is set like so - see: http://en.wikipedia.org/wiki/Kalman_filter#Example_application.2C_technical
+    P[0][1] = 0;
+    P[1][0] = 0;
+    P[1][1] = 0;
 };
 
 // The angle should be in degrees and the rate should be in degrees per second and the delta time in seconds
-float Kalman::get(float newX, float newRate, float dt) {
+double Kalman::get(double newX, double newRate, double dt) {
     // KasBot V2  -  Kalman filter module - http://www.x-firm.com/?page_id=145
     // Modified by Kristian Lauszus
     // See my blog post for more information: http://blog.tkjelectronics.dk/2012/09/a-practical-approach-to-kalman-filter-and-how-to-implement-it
@@ -54,23 +54,23 @@ float Kalman::get(float newX, float newRate, float dt) {
     // Discrete Kalman filter measurement update equations - Measurement Update ("Correct")
     // Calculate Kalman gain - Compute the Kalman gain
     /* Step 4 */
-    float S = P[0][0] + R_measure; // Estimate error
+    double S = P[0][0] + R_measure; // Estimate error
     /* Step 5 */
-    float K[2]; // Kalman gain - This is a 2x1 vector
+    double K[2]; // Kalman gain - This is a 2x1 vector
     K[0] = P[0][0] / S;
     K[1] = P[1][0] / S;
 
     // Calculate angle and bias - Update estimate with measurement zk (newAngle)
     /* Step 3 */
-    float y = newX - x; // Angle difference
+    double y = newX - x; // Angle difference
     /* Step 6 */
     x += K[0] * y;
     bias += K[1] * y;
 
     // Calculate estimation error covariance - Update the error covariance
     /* Step 7 */
-    float P00_temp = P[0][0];
-    float P01_temp = P[0][1];
+    double P00_temp = P[0][0];
+    double P01_temp = P[0][1];
 
     P[0][0] -= K[0] * P00_temp;
     P[0][1] -= K[0] * P01_temp;
@@ -80,16 +80,16 @@ float Kalman::get(float newX, float newRate, float dt) {
     return x;
 };
 
-void Kalman::set(float x) { this->x = x; }; // Used to set angle, this should be set as the starting angle
-float Kalman::getRate() { return this->rate; }; // Return the unbiased rate
+void Kalman::set(double x) { this->x = x; }; // Used to set angle, this should be set as the starting angle
+double Kalman::getRate() { return this->rate; }; // Return the unbiased rate
 
 /* These are used to tune the Kalman filter */
-void Kalman::setParameters(float Q, float Q_bias, float R_measure) { 
-	this->Q = Q; 
+void Kalman::setParameters(double Q, double Q_bias, double R_measure) {
+	this->Q = Q;
 	this->Q_bias = Q_bias;
 	this->R_measure = R_measure;
 };
 
-float Kalman::getQ() { return this->Q; };
-float Kalman::getQbias() { return this->Q_bias; };
-float Kalman::getRmeasure() { return this->R_measure; };
+double Kalman::getQ() { return this->Q; };
+double Kalman::getQbias() { return this->Q_bias; };
+double Kalman::getRmeasure() { return this->R_measure; };
