+++ libraries/KalmanFilter-master/Kalman.h
@@ -23,29 +23,29 @@ public:
     Kalman();
 
     // The angle should be in degrees and the rate should be in degrees per second and the delta time in seconds
-    float get(float newX, float newRate, float dt);
+    double get(double newX, double newRate, double dt);
 
-    void set(float x); // Used to set angle, this should be set as the starting angle
-    float getRate(); // Return the unbiased rate
+    void set(double x); // Used to set angle, this should be set as the starting angle
+    double getRate(); // Return the unbiased rate
 
     /* These are used to tune the Kalman filter */
-    void setParameters(float Q, float Q_bias, float R_measure);
+    void setParameters(double Q, double Q_bias, double R_measure);
 
-    float getQ();
-    float getQbias();
-    float getRmeasure();
+    double getQ();
+    double getQbias();
+    double getRmeasure();
 
 private:
     /* Kalman filter variables */
-    float Q; // Process noise variance for the accelerometer
-    float Q_bias; // Process noise variance for the gyro bias
-    float R_measure; // Measurement noise variance - this is actually the variance of the measurement noise
+    double Q; // Process noise variance for the accelerometer
+    double Q_bias; // Process noise variance for the gyro bias
+    double R_measure; // Measurement noise variance - this is actually the variance of the measurement noise
 
-    float x; // The angle calculated by the Kalman filter - part of the 2x1 state vector
-    float bias; // The gyro bias calculated by the Kalman filter - part of the 2x1 state vector
-    float rate; // Unbiased rate calculated from the rate and the calculated bias - you have to call getAngle to update the rate
+    double x; // The angle calculated by the Kalman filter - part of the 2x1 state vector
+    double bias; // The gyro bias calculated by the Kalman filter - part of the 2x1 state vector
+    double rate; // Unbiased rate calculated from the rate and the calculated bias - you have to call getAngle to update the rate
 
-    float P[2][2]; // Error covariance matrix - This is a 2x2 matrix
+    double P[2][2]; // Error covariance matrix - This is a 2x2 matrix
 };
 
 #endif
